refactor: centralize invite actions under [token] route
